USB: serial: xr: add support for multi-port XR21V141X variants
XR21V1412 and XR21V1414 have exactly the same interface, but
they support multiple 2 and 4 ports, respectively.

On such devices, the "CDC Union" field shows how they're
grouped, as can be seen on those lsusb -v outputs:

	https://linux-usb.vger.kernel.narkive.com/YaTYwHkM/usb-uart-device-from-exar-co-not-working-with-cdc-acm-but-usbserial
	https://jquery.developreference.com/article/22043997/udev+rule+with+bInterfaceNumber+doesn't+work+%5bclosed%5d

So, for instance, on XR21V1414, (0x04e2:0x1414), the 3rd
port is:

	CDC Union:
		bMasterInterface 4
		bSlaveInterface 5

	CDC Call Management:
		bmCapabilities 0x01
			call management
				bDataInterface 5

In other words, the control interface is an even number,
and the data interface is the next odd number.

The logic to get the proper register for an specific channel
came from this patch:

	https://lore.kernel.org/linux-usb/20180404070634.nhspvmxcjwfgjkcv@advantechmxl-desktop

Add support for them.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>